Fix bug with valid field not included in output
The problem is that go doesn't include 'valid' in the output if it set
to false, which appears to be by design, since false is the 'nil-value'
for a bool. Since that's an issue though as Alfred defaults to true for
a missing 'valid', this doesn't work.

There's two options here. Either make 'valid' a pointer allowing for all
three cases (true, false and nil) or make 'valid' not 'omitempty' but
default it to true in a constructor function. Since there's no way to
ensure however that a user actually uses that constructor (unless the type
is no longer exported, but that's also kinda weird), that would only
result in unexpected behavior. The only sane, albeit a bit weird, solution
is the pointer^^